Criado componente de redefinir senha.
Separado em 2 componentes o popup de edição de usuários.
Criado o componente userDadosPopup para fazer a renderização dos dados.
Criado o componente userResetPass para renderizar a redefinicao de senha
Criado a função para fazer a requisição para a API.
